internal/appsec: refactor IP related code
